refactor: move the modal layer into input/vi/
Vi mode is a specialization of VT input rather than a peer concern, so it nests
under input/ instead of sitting beside it. The include graph is what makes that
safe rather than merely tidy, and it runs one way:

  ViInputHandler includes InputHandler -- it IMPLEMENTS the interface input/ owns.
  Selector includes InputGenerator.
  Nothing in input/ includes anything in input/vi/, and a grep enforces it:

    grep '#include <vtbackend/input/vi/' src/vtbackend/input/*.{hpp,cpp}

HintModeHandler and JumpHistory name nothing from vi themselves, but vi is what
drives them -- ViInputHandler and ViCommands both do -- so they sit on the vi side
of the seam.

Selector comes here rather than to grid/, which is where a selection sounds like
it belongs. It includes InputGenerator and ViInputHandler and NOTHING from the
grid: it reaches cells only through its SelectionHelper dependency-injection seam,
over CellLocation. Its constructor takes a ViMode, and its four subclasses --
Linear, FullLine, Rectangular, WordWise -- are v, V and <C-V>.

This is the second and last two-level nesting in the module, the other being
graphics/regis/, and both are there for the same reason.
