+++ src/vtbackend/input/vi/HintModeHandler.hpp
