+++ src/vtbackend/input/vi/HintModeHandler_test.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/HintModeHandler.hpp>
 #include <vtbackend/core/FileUrl.hpp>
+#include <vtbackend/input/vi/HintModeHandler.hpp>
 
 #include <catch2/catch_test_macros.hpp>
