+++ src/vtbackend/input/vi/JumpHistory.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/JumpHistory.hpp>
+#include <vtbackend/input/vi/JumpHistory.hpp>
 
 namespace vtbackend
 {
